# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar-IndigoPay API
  version: 1.0.0
  description: |
    OpenAPI 3.0 documentation for the backend REST API used by Stellar-IndigoPay.
    The API exposes project discovery, donations, profiles, leaderboard, notifications,
    ratings, updates, jobs, stats, and admin management endpoints.
  contact:
    name: Stellar-IndigoPay Team
    url: https://github.com/Stellar-IndigoPay/Stellar-IndigoPay
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development server  
  - url: https://api.stellarindigopay.com
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      description: Project identifier.
      schema:
        type: string
...
    LimitQuery:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CursorQuery:
      name: cursor
      in: query
      schema:
        type: string
...
  responses:
    BadRequest:
      description: The request payload or query parameters are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication is required or the provided token is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimitExceeded:
      description: The request has been rate-limited. All endpoints carry rate-limit response headers (X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset) and return 429 with a Retry-After header when the limit is exceeded. See the API documentation for per-endpoint tiers.
      headers:
        Retry-After:
          description: Number of seconds to wait before retrying.
          schema:
            type: integer
        X-RateLimit-Limit:
          description: Maximum requests allowed (or bucket capacity for token-bucket endpoints).
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Requests remaining in the current window (or tokens left in the bucket).
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Epoch seconds when the limit resets (or next token becomes available).
          schema:
            type: integer
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Too many requests — Try again later.
              retryAfter:
                type: integer
                description: Epoch seconds after which a retry may succeed.
                example: 1742169602
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid request payload
      required:
        - error
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        category:
          type: string
        location:
          type: string
        walletAddress:
          type: string
        goalXLM:
          type: string
        raisedXLM:
          type: string
        donorCount:
          type: integer
        co2OffsetKg:
          type: integer
        status:
          type: string
          enum: [active, completed, paused, rejected]
        rejectionReason:
          type: string
          nullable: true
        verified:
          type: boolean
        onChainVerified:
          type: boolean
        tags:
          type: array
          items:
            type: string
        aiSummary:
          type: string
          nullable: true
        aiSummaryGeneratedAt:
          type: string
          format: date-time
          nullable: true
        aiSummaryModel:
          type: string
          nullable: true
        aiSummarySourceHash:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateProjectRequest:
      type: object
      required: [name, description, location, category, wallet_address]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 120
        description:
          type: string
          minLength: 10
          maxLength: 5000
        location:
          type: string
          minLength: 2
          maxLength: 200
        category:
          type: string
          enum:
            [
              Reforestation,
              Solar Energy,
              Ocean Conservation,
              Clean Water,
              Wildlife Protection,
              Carbon Capture,
              Wind Energy,
              Sustainable Agriculture,
              Other,
            ]
        wallet_address:
          type: string
          pattern: "^G[A-Z0-9]{55}$"
        goal_xlm:
          type: number
          default: 0
        tags:
          type: array
          items:
            type: string
...
    Attestation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        onChainId:
          type: integer
          nullable: true
        sourceChain:
          type: string
          description: Source chain slug (e.g. "ethereum", "polygon", "arbitrum", "base").
          example: ethereum
        sourceTxHash:
          type: string
          description: Transaction hash on the source chain.
          example: "0xabc123…"
        donorAddress:
          type: string
          pattern: "^G[A-Z0-9]{55}$"
        projectId:
          type: string
          format: uuid
          nullable: true
        amountUsd:
          type: string
          description: USD-equivalent value, 6 decimals (USDC convention).
          example: "10.000000"
        amountXlm:
          type: string
          description: XLM-equivalent value, 7 decimals (stroops stored as decimal).
          example: "80.0000000"
        status:
          type: string
          enum: [pending, verified, revoked]
        messageHash:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        verifiedAt:
          type: string
          format: date-time
          nullable: true
    AttestationStats:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        verified:
          type: integer
        revoked:
          type: integer
        byChain:
          type: array
          items:
            type: object
            properties:
              sourceChain:
                type: string
              count:
                type: integer
...
    RecordAttestationRequest:
      type: object
      required:
        - source_chain
        - source_tx_hash
        - donor_address
        - project_id
        - amount_usd
        - amount_xlm
        - on_chain_id
      properties:
        source_chain:
          type: string
        source_tx_hash:
          type: string
        donor_address:
          type: string
          pattern: "^G[A-Z0-9]{55}$"
        project_id:
          type: string
          format: uuid
        amount_usd:
          type: number
          description: USD-equivalent value, 6 decimals.
        amount_xlm:
          type: number
          description: XLM-equivalent value, 7 decimals.
        message_hash:
          type: integer
        on_chain_id:
          type: integer
          description: Soroban-assigned monotonic id returned by the relayer.
        status:
          type: string
          enum: [pending, verified, revoked]
          default: pending
        recorded_by:
          type: string
...
paths:
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List projects
      operationId: listProjects
      description: Retrieve a paginated list of projects with optional filters for category, status, verification, and search.
      parameters:
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/CursorQuery"
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, paused]
        - name: verified
          in: query
          schema:
            type: string
            enum: ["true"]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated project list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
                  next_cursor:
                    type: string
                    nullable: true
                  has_more:
                    type: boolean
    post:
      tags: [Projects]
      summary: Create a project
      operationId: createProject
      description: Register a new climate project on the platform with name, description, location, category, and wallet address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
...
  /api/projects/{id}:
    get:
      tags: [Projects]
      summary: Get a project by id
      operationId: getProjectById
      description: Retrieve detailed information about a specific project by its unique identifier.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/attestations:
    get:
      tags: [Attestations]
      operationId: getAttestations
      summary: Get attestation roll-up stats
      description: Returns platform-wide counts (total / pending / verified / revoked) plus per-source-chain breakdown (issue #125).
      security: []
      responses:
        "200":
          description: Roll-up stats.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/AttestationStats"
    post:
      tags: [Attestations]
      operationId: recordAttestation
      summary: Record a new cross-chain attestation
      description: Called by the relayer after source-chain finality. Requires a signed proof in either the `x-attestation-signature` / `x-attestation-timestamp` headers or the legacy `x_proof_signature` / `x_proof_timestamp` aliases.
      security: []
      parameters:
        - name: x-attestation-signature
          in: header
          required: true
          schema:
            type: string
        - name: x-attestation-timestamp
          in: header
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordAttestationRequest"
      responses:
        "201":
          description: Attestation created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  created:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Attestation"
        "200":
          description: Attestation already existed; idempotent response.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  created:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Attestation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
...
  /api/attestations/{id}:
    get:
      tags: [Attestations]
      operationId: getAttestationById
      summary: Get an attestation by backend UUID
      description: Retrieve a single attestation by its backend UUID identifier.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The matching attestation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Attestation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...